test/Task_repository/app/requirements.txt
pymongo
paho-mqtt==2.1.0
PyYAML==6.0.1